+++ src/components/Search/common.js
